Move files into suitable folders. Need check more.
