+++ Learning System/Learning System/Controls/EditQuestionForm.cs
@@ -1,5 +1,6 @@
-﻿using Learning_System.Class;
-using Learning_System.ExternalClass;
+﻿using Learning_System.ProcessingClasses;
+using Learning_System.Modals;
+using Learning_System.ProcessingClasses;
 using Newtonsoft.Json.Linq;
 using System.CodeDom;
 using System.Data;
